requirements.txt
akshare
pandas
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
tqdm
openpyxl